Plant/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.32
pydantic>=2.8.2
python-multipart>=0.0.9
apscheduler>=3.10.4
numpy
pyyaml>=6.0.2
torch
torchvision
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt<4.0.0
thop
tqdm
PyYAML
matplotlib
albumentations
ultralytics
opencv-python==4.12.0.88 
opencv-python-headless==4.12.0.88
